Additional Sample buffer tests exposed a truncation issue in SampleProcessor.

+++ AudioWorks/src/AudioWorks.Extensibility/SampleProcessor.cs
@@ -117,7 +117,7 @@ namespace AudioWorks.Extensibility
                 for (var sampleIndex = sourceVectors.Length * Vector<float>.Count;
                     sampleIndex < source.Length;
                     sampleIndex++)
-                    destination[sampleIndex] = (short) Math.Min(source[sampleIndex] * multiplier, max);
+                    destination[sampleIndex] = (int) Math.Min(source[sampleIndex] * multiplier, max);
             }
             else
             {
